move account into individual directory
